Day 1 and 2 of code with python done!
